refactor: move constnats to common dir
